requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.7
httpx==0.27.0
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
